# web/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Combina API",
    "version": "1.0.0",
    "description": "Lottery combination generation and evaluation API."
  },
  "servers": [
    { "url": "http://localhost:3000" }
  ],
  "paths": {
...
    "/combinations/generate-from": {
      "post": {
        "summary": "Generate new games combined with existing imported games",
        "operationId": "generateFrom",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/GenerateFromInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created combination containing imported + generated games",
            "headers": {
              "Location": { "schema": { "type": "string" }, "description": "ID of the new combination" }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Lotto" }
              }
            }
          },
          "400": { "description": "Missing or invalid fields", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "409": { "description": "Combination already exists", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "500": { "description": "Internal server error", "content": { "text/plain": { "schema": { "type": "string" } } } }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "GameCombo": {
        "type": "object",
        "properties": {
          "combination": {
            "type": "array",
            "items": { "type": "array", "items": { "type": "integer" } },
            "description": "List of games, each game is a list of numbers"
          },
          "rows": { "type": "integer", "description": "Number of games" },
          "cols": { "type": "integer", "description": "Numbers per game (generated games)" }
        }
      },
      "Lotto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "numbers": { "$ref": "#/components/schemas/GameCombo" },
          "game_type": { "type": "string", "example": "Lotofacil" },
          "created_on": { "type": "string", "format": "date-time" },
          "alias": { "type": "string", "example": "meus jogos" }
        }
      },
...
      "GenerateFromInput": {
        "type": "object",
        "required": ["game_type", "num_games", "num_each"],
        "properties": {
          "game_type": { "type": "string", "example": "Lotofacil" },
          "num_games": { "type": "integer", "example": 3 },
          "num_each": { "type": "integer", "example": 15 },
          "existing_games": {
            "type": "array",
            "items": { "type": "array", "items": { "type": "integer" } },
            "description": "Games previously imported; stored alongside the generated ones"
          },
          "fixed_numbers": {
            "type": "array",
            "items": { "type": "integer" },
            "description": "Numbers that must appear in every game"
          },
          "favored": {
            "type": "array",
            "items": { "type": "integer" },
            "description": "Numbers that are favored (7:3 probability ratio) when non-empty"
          },
          "disfavored": {
            "type": "array",
            "items": { "type": "integer" },
            "description": "Numbers that appear less frequently (0.1x weight)"
          },
          "excluded_numbers": {
            "type": "array",
            "items": { "type": "integer" },
            "description": "Numbers that must never appear in any generated game"
          },
          "alias": { "type": "string", "example": "meus jogos" }
        }
      },
...
    }
  }
}